debugging awslambda test case for public alb check
